kitty

+++ .config/fish/functions/config-test.fish
@@ -7,17 +7,17 @@ function config-test
 
     set argument $argv[1]
 
-    switch $argument
-        case nvim hypr fish kitty starship
-            # Remove existing configuration
-            rm -rf "$HOME/.config/$argument"
-            
-            # Create symlink to ~/.dotfiles/.config/$argument
-            ln -s ~/.dotfiles/.config/$argument ~/.config/$argument
-            
-            echo "Configuration for $argument removed and symlink created."
-        case '*'
-            echo "Invalid argument. Supported values: nvim, hypr, fish"
-            return 1
+    if test "$argument" = "nvim" -o "$argument" = "hypr" -o "$argument" = "fish" -o "$argument" = "kitty" -o "$argument" = "starship"
+        # Remove existing configuration
+        rm -rf "$HOME/.config/$argument"
+
+        # Create symlink to ~/.dotfiles/.config/$argument
+        ln -s ~/.dotfiles/.config/$argument ~/.config/$argument
+
+        echo "Configuration for $argument removed and symlink created."
+    else
+        echo "Invalid argument. Supported values: nvim, hypr, fish, kitty, starship"
+        return 1
     end
 end
+
